admin-api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0
dj-database-url==2.1.0
